# [SPIKE] MSMQ use raw byte segments from the message without copying