--- a/src/NServiceBus.Core/Recoverability/DelayedRetryExecutor.cs
+++ b/src/NServiceBus.Core/Recoverability/DelayedRetryExecutor.cs
@@ -20,8 +20,7 @@
 
         public async Task<int> Retry(IncomingMessage message, TimeSpan delay, TransportTransaction transportTransaction)
         {
-            var outgoingMessage = new OutgoingMessage(message.MessageId, new Dictionary<string, string>(message.Headers), message.Body);
-
+            var outgoingMessage = message.ToOutgoingMessage();
             var currentDelayedRetriesAttempt = message.GetDelayedDeliveriesPerformed() + 1;
 
             outgoingMessage.SetCurrentDelayedDeliveries(currentDelayedRetriesAttempt);