--- a/src/NServiceBus.Core/Reliability/Outbox/TransportOperation.cs
+++ b/src/NServiceBus.Core/Reliability/Outbox/TransportOperation.cs
@@ -1,5 +1,6 @@
 ﻿namespace NServiceBus.Outbox
 {
+    using System;
     using System.Collections.Generic;
 
     /// <summary>
@@ -26,6 +27,24 @@
         }
 
         /// <summary>
+        /// Creates a new instance of a <see cref="TransportOperation" />.
+        /// </summary>
+        /// <param name="messageId">The identifier of the outgoing message.</param>
+        /// <param name="options">The sending options.</param>
+        /// <param name="body">The message body.</param>
+        /// <param name="headers">The message headers.</param>
+        /// .
+        public TransportOperation(string messageId, Dictionary<string, string> options, ArraySegment<byte> body, Dictionary<string, string> headers)
+        {
+            Guard.AgainstNullAndEmpty(nameof(messageId), messageId);
+
+            MessageId = messageId;
+            Options = options;
+            BodySegment = body;
+            Headers = headers;
+        }
+
+        /// <summary>
         /// Gets the identifier of the outgoing message.
         /// </summary>
         public string MessageId { get; private set; }
@@ -39,6 +58,11 @@
         /// Gets a byte array to the body content of the outgoing message.
         /// </summary>
         public byte[] Body { get; private set; }
+
+        /// <summary>
+        /// Gets a byte array to the body content of the outgoing message.
+        /// </summary>
+        public ArraySegment<byte> BodySegment { get; private set; }
 
         /// <summary>
         /// Gets outgoing message headers.