--- a/src/NServiceBus.Core/Transports/MessageContext.cs
+++ b/src/NServiceBus.Core/Transports/MessageContext.cs
@@ -1,5 +1,6 @@
 ﻿namespace NServiceBus.Transport
 {
+    using System;
     using System.Collections.Generic;
     using System.Threading;
     using Extensibility;
@@ -40,6 +41,36 @@
         }
 
         /// <summary>
+        /// Initializes the context.
+        /// </summary>
+        /// <param name="messageId">Native message id.</param>
+        /// <param name="headers">The message headers.</param>
+        /// <param name="body">The message body.</param>
+        /// <param name="transportTransaction">Transaction (along with connection if applicable) used to receive the message.</param>
+        /// <param name="receiveCancellationTokenSource">
+        /// Allows the pipeline to flag that it has been aborted and the receive operation should be rolled back.
+        /// It also allows the transport to communicate to the pipeline to abort if possible. Transports should check if the token
+        /// has been aborted after invoking the pipeline and roll back the message accordingly.
+        /// </param>
+        /// <param name="context">Any context that the transport wants to be available on the pipeline.</param>
+        public MessageContext(string messageId, Dictionary<string, string> headers, ArraySegment<byte> body, TransportTransaction transportTransaction, CancellationTokenSource receiveCancellationTokenSource, ContextBag context)
+        {
+            Guard.AgainstNullAndEmpty(nameof(messageId), messageId);
+            Guard.AgainstNull(nameof(body), body);
+            Guard.AgainstNull(nameof(headers), headers);
+            Guard.AgainstNull(nameof(transportTransaction), transportTransaction);
+            Guard.AgainstNull(nameof(receiveCancellationTokenSource), receiveCancellationTokenSource);
+            Guard.AgainstNull(nameof(context), context);
+
+            Headers = headers;
+            BodySegment = body;
+            MessageId = messageId;
+            Context = context;
+            TransportTransaction = transportTransaction;
+            ReceiveCancellationTokenSource = receiveCancellationTokenSource;
+        }
+
+        /// <summary>
         /// The native id of the message.
         /// </summary>
         public string MessageId { get; }
@@ -53,6 +84,11 @@
         /// The message body.
         /// </summary>
         public byte[] Body { get; }
+
+        /// <summary>
+        /// The message body segment.
+        /// </summary>
+        public ArraySegment<byte> BodySegment { get; }
 
         /// <summary>
         /// Transaction (along with connection if applicable) used to receive the message.