--- a/src/NServiceBus.Core/Transports/Msmq/ReceiveOnlyNativeTransactionStrategy.cs
+++ b/src/NServiceBus.Core/Transports/Msmq/ReceiveOnlyNativeTransactionStrategy.cs
@@ -82,7 +82,7 @@
                 {
                     using (var bodyStream = message.BodyStream)
                     {
-                        var shouldAbortMessageProcessing = await TryProcessMessage(message.Id, headers, bodyStream, transportTransaction).ConfigureAwait(false);
+                        var shouldAbortMessageProcessing = await TryProcessMessage(message, headers, bodyStream, transportTransaction).ConfigureAwait(false);
 
                         if (shouldAbortMessageProcessing)
                         {