… src/NServiceBus.Core/Transports/Msmq/MsmqUtilities.cs
… src/NServiceBus.Core/Transports/Msmq/ReceiveStrategy.cs
… src/NServiceBus.Core/Transports/OutgoingMessage.cs
… src/NServiceBus.Core/Unicast/IncomingMessageOperations.cs
… src/NServiceBus.Core/Unicast/MessageOperationsInvokeHandlerContext.cs