@danielmarbach  and I where discussing the buffering PR #4227 and decided that it is actually unnecessary to introduce buffering and all the cruft that comes with it since MSMQ will always allocate a byte array per message anyway. With this PR we want to explore the pros and cons of bypassing the public API of memory stream and directly access the byte array which is held internally. Our main argument was that `MemoryStream` and Msmq is unlikely to change that behavior (we also checked memory stream imp in .NET core). Since MSMQ will remain a full framework thing only we can assume to always have reflection EMIT as well as the stable behavior of MSMQ and the memory stream.