@timbussmann 2016-10-19
Ok, while I think this is a cool solution I think we shouldn't do this.

> Our main argument was that MemoryStream and Msmq is unlikely to change that behavior

Depending on internal implementation details to get a little bit of performance seems like a really bad tradeoff for a framework. I'm not going to argue further than that as of now. How would you react when users make use of internal implementation details to improve NSB peformance? Would you consider this a good approach? Imho, unlikeliness isn't any sort of argument to bring up in a discussion like this.