@danielmarbach 2016-10-19
I think the question we have to ask ourself is:

Do we want high perf and low memory pressure and how far are we willing to go.