@danielmarbach 2016-10-19
And before we even talk about what we do here we need to gather some facts, that is the purpose of this spike to answer questions like:

- How much throughput do we get
- How does it relax memory pressure / does it even
- What are the downsides of it
- How does this impact mutators / does it even
- How does taken the raw byte array and potentially modifying it affect transactions with MSQM
- ... more

If we know answers to above question then we can have a meaningful discussion whether it makes sense to do the nasty trickery outlined in the PR.