@timbussmann 2016-10-19
Feel free to go ahead an gather the data, surely will be interesting to know.

> Do we want high perf and low memory pressure and how far are we willing to go.

I don't think hacking into internals is a valid place to go, no matter how willing you are. Are you willing to rob a bank to get rich? Sorry for that bad analogy ;)