@danielmarbach 2016-10-19
The problem is that when you want to go high perf you end up doing "unpure" things like robbing a bank (aka memory pinning ;) ). It is not black and white, it is more shades of grey