@timbussmann 2016-10-19
that may be an acceptable approach for your own application where you have to justify the risks you take to yourself but imho not for a framework running in uncontrollable environments. You force users to take a risk you're willing to take. The question is: are they willing to rob a bank with you? Spoiler alert: I'm not.