@Scooletz 2016-10-19
@timbussmann Although I see a point about not 

> Depending on internal implementation details to get a little bit of performance 

although I don't agree with 

> not for a framework running in uncontrollable environments

as this is controllable. I think that it would be worth it to run a performance benchmark and gather data, then decide whether or not consider this spike as a real PR.