@SzymonPobiega 2016-10-19
:popcorn: I like the discussion. Keep it going. If I understand correctly we are talking about avoiding (re)allocation of entire message body. If I can express my feelings, doing this nasty thing in a single defined place potentially has far greater impact on memory pressure than all my `IEnumerables` I used in last two years. And from the maintainability point of view it seems to me much less harmful than turning all LINQ usages into for loops.