@timbussmann 2016-11-07
@Scooletz @danielmarbach any updates on meassurements?