@Scooletz 2016-11-07
Not yet @timbussmann I was silently waiting for @danielmarbach but if you had a BenchmarkDotNet scenario I could give it a run. If you don't have something like this, I can prepare one this week and then, you could validate it maybe?