@danielmarbach 2016-11-23
It was pretty difficult to profile that one properly. Profiling it with dotMemory, Ants, YourKit did not show any significant improvements in terms of perf and GC pauses. I used my SwissChocolate.Power samples with receiving 10K messages each. 

## Before

<img width="804" alt="withoutrawaccess" src="https://cloud.githubusercontent.com/assets/174258/20564777/0b1846fe-b18f-11e6-953c-883c01b34d47.PNG">

![image](https://cloud.githubusercontent.com/assets/174258/20564909/89cfc6f2-b18f-11e6-8e84-b9a7e8d0ffce.png)

## After

<img width="776" alt="withrawaccess2" src="https://cloud.githubusercontent.com/assets/174258/20564801/1ef38a08-b18f-11e6-885d-267d858f0e9a.PNG">

![image](https://cloud.githubusercontent.com/assets/174258/20564848/579c9b24-b18f-11e6-9bb0-cc240196777d.png)


## Micro Benchmark results

[Benchmark used](https://github.com/danielmarbach/MicroBenchmarks/commit/4d75aadb4a59755af4c389a3aa9c287f2c4e3b8d)

```ini

Host Process Environment Information:
BenchmarkDotNet=v0.9.8.0
OS=Microsoft Windows NT 6.2.9200.0
Processor=Intel(R) Core(TM) i7-4980HQ CPU 2.80GHz, ProcessorCount=8
Frequency=2728067 ticks, Resolution=366.5599 ns, Timer=TSC
CLR=MS.NET 4.0.30319.42000, Arch=32-bit RELEASE
GC=Concurrent Workstation
JitModules=clrjit-v4.6.1586.0

Type=MsmqBodyStream  Mode=SingleRun  GarbageCollection=Concurrent Workstation  
LaunchCount=1  WarmupCount=1  TargetCount=1  

```
              Method | Platform | BodySize | NumberOfMessages |      Median |    StdDev | Scaled |  Gen 0 |  Gen 1 | Gen 2 | Bytes Allocated/Op |
-------------------- |--------- |--------- |----------------- |------------ |---------- |------- |------- |------- |------ |------------------- |
 **V6_BodyStreamAccess** |      **X64** |      **256** |             **1000** |  **26.0774 ms** | **0.0000 ms** |   **1.00** |   **2.00** |   **2.00** |     **-** |      **22'851'421.33** |
 V6_BodyDirectAccess |      X64 |      256 |             1000 |  25.9781 ms | 0.0000 ms |   1.00 |   2.00 |   2.00 |     - |      22'274'933.33 |
 V6_BodyStreamAccess |      X86 |      256 |             1000 |  28.9905 ms | 0.0000 ms |   1.00 |   3.00 |   5.00 |     - |      16'773'670.67 |
 V6_BodyDirectAccess |      X86 |      256 |             1000 |  25.0335 ms | 0.0000 ms |   0.86 |   2.00 |   3.00 |  1.00 |      16'291'244.00 |
 **V6_BodyStreamAccess** |      **X64** |      **256** |            **10000** | **269.5630 ms** | **0.0000 ms** |   **1.00** |  **50.00** |  **40.00** |  **5.00** |     **228'116'976.00** |
 V6_BodyDirectAccess |      X64 |      256 |            10000 | 284.7056 ms | 0.0000 ms |   1.06 |  61.00 |  25.00 |  7.00 |     222'296'130.67 |
 V6_BodyStreamAccess |      X86 |      256 |            10000 | 295.8197 ms | 0.0000 ms |   1.00 |   5.00 |  78.00 | 10.00 |     167'446'432.00 |
 V6_BodyDirectAccess |      X86 |      256 |            10000 | 306.4972 ms | 0.0000 ms |   1.04 |   9.00 |  65.00 | 12.00 |     162'789'292.00 |
 **V6_BodyStreamAccess** |      **X64** |      **256** |            **20000** | **534.1269 ms** | **0.0000 ms** |   **1.00** | **121.00** |  **71.00** | **10.00** |     **456'198'754.67** |
 V6_BodyDirectAccess |      X64 |      256 |            20000 | 557.0028 ms | 0.0000 ms |   1.04 | 121.00 |  70.00 | 14.00 |     444'635'181.33 |
 V6_BodyStreamAccess |      X86 |      256 |            20000 | 568.7943 ms | 0.0000 ms |   1.00 |   7.00 | 169.00 | 13.00 |     334'882'674.67 |
 V6_BodyDirectAccess |      X86 |      256 |            20000 | 588.1311 ms | 0.0000 ms |   1.03 |  12.00 | 143.00 | 13.00 |     325'570'621.33 |
 **V6_BodyStreamAccess** |      **X64** |      **512** |             **1000** |  **25.6332 ms** | **0.0000 ms** |   **1.00** |   **2.00** |   **2.00** |     **-** |      **23'914'674.67** |
 V6_BodyDirectAccess |      X64 |      512 |             1000 |  25.2149 ms | 0.0000 ms |   0.98 |   2.00 |   2.00 |     - |      22'996'416.00 |
 V6_BodyStreamAccess |      X86 |      512 |             1000 |  28.4040 ms | 0.0000 ms |   1.00 |   3.00 |   5.00 |  1.00 |      17'824'353.33 |
 V6_BodyDirectAccess |      X86 |      512 |             1000 |  24.0705 ms | 0.0000 ms |   0.85 |   2.00 |   3.00 |  1.00 |      17'003'797.33 |
 **V6_BodyStreamAccess** |      **X64** |      **512** |            **10000** | **277.9609 ms** | **0.0000 ms** |   **1.00** |  **55.00** |  **38.00** |  **8.00** |     **238'892'501.33** |
 V6_BodyDirectAccess |      X64 |      512 |            10000 | 282.4648 ms | 0.0000 ms |   1.02 |  62.00 |  28.00 |  8.00 |     229'469'840.00 |
 V6_BodyStreamAccess |      X86 |      512 |            10000 | 298.9021 ms | 0.0000 ms |   1.00 |   7.00 |  81.00 | 12.00 |     178'201'774.67 |
 V6_BodyDirectAccess |      X86 |      512 |            10000 | 306.4148 ms | 0.0000 ms |   1.03 |   8.00 |  68.00 | 12.00 |     169'975'750.67 |
 **V6_BodyStreamAccess** |      **X64** |      **512** |            **20000** | **553.0462 ms** | **0.0000 ms** |   **1.00** | **124.00** |  **80.00** | **12.00** |     **477'727'800.00** |
 V6_BodyDirectAccess |      X64 |      512 |            20000 | 559.2931 ms | 0.0000 ms |   1.01 | 112.00 |  85.00 | 12.00 |     458'976'272.00 |
 V6_BodyStreamAccess |      X86 |      512 |            20000 | 592.0155 ms | 0.0000 ms |   1.00 |  11.00 | 177.00 | 13.00 |     356'374'518.67 |
 V6_BodyDirectAccess |      X86 |      512 |            20000 | 598.9398 ms | 0.0000 ms |   1.01 |  11.00 | 152.00 | 12.00 |     339'901'457.33 |
 **V6_BodyStreamAccess** |      **X64** |     **1024** |             **1000** |  **25.9796 ms** | **0.0000 ms** |   **1.00** |   **2.00** |   **2.00** |     **-** |      **26'158'560.00** |
 V6_BodyDirectAccess |      X64 |     1024 |             1000 |  26.3190 ms | 0.0000 ms |   1.01 |   2.00 |   2.00 |     - |      24'494'629.33 |
 V6_BodyStreamAccess |      X86 |     1024 |             1000 |  27.0741 ms | 0.0000 ms |   1.00 |   2.00 |   4.00 |  1.00 |      20'059'356.00 |
 V6_BodyDirectAccess |      X86 |     1024 |             1000 |  25.4077 ms | 0.0000 ms |   0.94 |   2.00 |   4.00 |  1.00 |      18'501'905.33 |
 **V6_BodyStreamAccess** |      **X64** |     **1024** |            **10000** | **267.1459 ms** | **0.0000 ms** |   **1.00** |  **71.00** |  **34.00** |  **9.00** |     **261'440'826.67** |
 V6_BodyDirectAccess |      X64 |     1024 |            10000 | 290.8272 ms | 0.0000 ms |   1.09 |  61.00 |  37.00 |  8.00 |     244'528'762.67 |
 V6_BodyStreamAccess |      X86 |     1024 |            10000 | 312.0744 ms | 0.0000 ms |   1.00 |   8.00 |  80.00 | 12.00 |     200'667'629.33 |
 V6_BodyDirectAccess |      X86 |     1024 |            10000 | 313.1969 ms | 0.0000 ms |   1.00 |   9.00 |  76.00 | 12.00 |     184'934'732.00 |
 **V6_BodyStreamAccess** |      **X64** |     **1024** |            **20000** | **598.3827 ms** | **0.0000 ms** |   **1.00** | **139.00** |  **86.00** | **12.00** |     **522'822'186.67** |
 V6_BodyDirectAccess |      X64 |     1024 |            20000 | 572.5105 ms | 0.0000 ms |   0.96 | 129.00 |  80.00 | 11.00 |     488'984'424.00 |
 V6_BodyStreamAccess |      X86 |     1024 |            20000 | 616.4841 ms | 0.0000 ms |   1.00 |  12.00 | 181.00 | 15.00 |     401'346'060.00 |
 V6_BodyDirectAccess |      X86 |     1024 |            20000 | 601.9442 ms | 0.0000 ms |   0.98 |  12.00 | 164.00 | 15.00 |     369'822'176.00 |